Refactored samplefilelist into separate components for sequencing objects and genome assemblies
